Improve event posts.new
If I want to catch only new posts (not change) for example for notifications, I don't think it is reasonable to call two `posts.submit` and` posts.new` events, even if this data is available during `posts.new`. There is no need to ignore the post data and then retrieve the same data by another extra query according to `insert_id` and` posttype`.